move XML beans to one class
